Fetch mainstream browse links from rummager
Rummager contains the links for mainstream browse pages. We should use
those to render the pages instead of querying the content-api.

This commit removes collections' dependency on content-api.
