+++ app/models/topic.rb
@@ -57,7 +57,7 @@ class Topic
   end
 
   def changed_documents
-    ChangedDocuments.new(slug, @pagination_options)
+    ChangedDocuments.new("specialist_sector", slug, @pagination_options)
   end
 
   def slug
