+++ test/models/topic_test.rb
@@ -114,13 +114,13 @@ describe Topic do
   end
 
   describe "lists" do
-    it "should pass the slug of the topic when constructing groups" do
+    it "passes the slug of the topic when constructing groups" do
       ListSet.expects(:new).with("specialist_sector", "business-tax/paye", anything()).returns(:a_lists_instance)
 
       assert_equal :a_lists_instance, @topic.lists
     end
 
-    it "should pass the groups data when constructing" do
+    it "passes the groups data when constructing" do
       ListSet.expects(:new).with(anything(), anything(), :some_data).returns(:a_lists_instance)
       @api_data["details"]["groups"] = :some_data
 
@@ -133,13 +133,13 @@ describe Topic do
       @topic = Topic.new(@api_data, {:foo => "bar"})
     end
 
-    it "should pass the contentapi slug of the topic when constructing changed_documents" do
-      Topic::ChangedDocuments.expects(:new).with("business-tax/paye", anything()).returns(:an_instance)
+    it "passes the contentapi slug of the topic when constructing changed_documents" do
+      Topic::ChangedDocuments.expects(:new).with("specialist_sector", "business-tax/paye", anything()).returns(:an_instance)
       assert_equal :an_instance, @topic.changed_documents
     end
 
-    it "should pass the pagination options when constructing changed_documents" do
-      Topic::ChangedDocuments.expects(:new).with(anything(), :foo => "bar").returns(:an_instance)
+    it "passes the pagination options when constructing changed_documents" do
+      Topic::ChangedDocuments.expects(:new).with("specialist_sector", anything(), :foo => "bar").returns(:an_instance)
       assert_equal :an_instance, @topic.changed_documents
     end
   end
