+++ test/test_helper.rb
@@ -20,12 +20,10 @@ WebMock.disable_net_connect!
 
 Dir[Rails.root.join('test/support/*.rb')].each { |f| require f }
 
-require 'gds_api/test_helpers/content_api'
 require 'gds_api/test_helpers/content_store'
 require 'gds_api/test_helpers/rummager'
 
 class ActiveSupport::TestCase
-  include GdsApi::TestHelpers::ContentApi
   include GdsApi::TestHelpers::ContentStore
   include GdsApi::TestHelpers::Rummager
